feature: process_with_memory
Added a previous_history attribute to many functions in an attempt to
build a pipeline with memory

The original idea was to limit overall slope and intercept changes to a
given percentage change on each, so that "jitteryness" would be reduced
in the extrapolated lane lines.

this became a neccessity, as the challenge video has frames where
hough lines drop off completely, requiring a knowledge of the previous
lines to be considered or it'll crash (we attempt to fit 0 points)

this is, at the moment, however, not working. I'm seeing the behavior
be WORSE than just allowing frame by frame change, which I don't
understand at all.